requirements.txt
streamlit
PyMuPDF
pandas
email-validator
pytesseract
pillow
python-docx
requests
beautifulsoup4
